@@ -8,19 +8,19 @@
     "npm": "9.8.1"
   },
   "dependencies": {
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^5.3.2",
     "clarifai": "^2.9.1",
-    "filestack-js": "^3.7.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-ga": "^2.6.0",
-    "react-remove-scroll": "^2.2.0",
-    "react-router-dom": "^5.2.0",
+    "filestack-js": "^3.27.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-remove-scroll": "^2.5.6",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
-    "react-tilt": "^0.1.4",
-    "react-tsparticles": "^2.3.3",
-    "reactstrap": "^8.0.1",
-    "tsparticles": "^2.3.4"
+    "react-tilt": "^1.0.2",
+    "react-tsparticles": "^2.12.2",
+    "reactstrap": "^9.2.0",
+    "tsparticles": "^2.12.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -29,12 +29,20 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app"
+    ]
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
